backend/requirements.txt
# Core Flask
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-CORS==4.0.1

# Authentication & Security
Flask-JWT-Extended==4.6.0
bcrypt==4.1.3
python-dotenv==1.0.1
Flask-Mail==0.10.0

# Database
SQLAlchemy==2.0.36
pg8000==1.31.2          # Pure-Python PostgreSQL driver (works on Python 3.13)

# PDF Generation
reportlab==4.2.2

# AI / LLM
openai==1.35.3
google-generativeai==0.8.3

# HTTP & Utils
requests==2.32.3
gunicorn==22.0.0

# WSGI server
Werkzeug==3.0.4
